requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.4
aiohttp-retry==2.9.1
aiosignal==1.3.2
attrs==25.3.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
Flask==3.1.1
frozenlist==1.6.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.4.4
propcache==0.3.1
PyJWT==2.10.1
python-dotenv==1.1.0
requests==2.32.3
twilio==9.6.2
urllib3==2.4.0
Werkzeug==3.1.3
yarl==1.20.0
